requirements.txt
Flask==0.10.1
Flask-Admin==1.3.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.1
Flask-Security==1.7.5
Flask-Social==1.6.2
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.10
WTForms==2.1
Werkzeug==0.11.3
blinker==1.4
facebook-sdk==1.0.0-alpha
httplib2==0.9.2
itsdangerous==0.24
oauth2==1.9.0.post1
oauthlib==1.0.3
passlib==1.6.5
python-twitter==2.2
requests==2.9.1
requests-oauthlib==0.6.0
webassets==0.11.1
wsgiref==0.1.2
